Arregla peligroso error con la template, y actualiza LDC

+++ template.cpp
@@ -3,7 +3,7 @@ using namespace std;
 #define forr(i, a, b) for(int i = (a); i < (int) (b); i++)
 #define forn(i, n) forr(i, 0, n)
 #define forall(it, v) for(auto it = v.begin(); it != v.end(); ++it)
-#define dforn(i, n) for(int i = n - 1; i >= 0; i--)
+#define dforn(i, n) for(int i = ((int) n) - 1; i >= 0; i--)
 #define db(v) cerr << #v << " = " << v << endl
 #define pb push_back
 typedef long long ll;
